- bulk upload
git-svn-id: file:///Users/jan/Dropbox/svn/FlexyAdmin/trunk@435 0c02b2c5-a8bd-4e35-85b2-cf8cacb96951

+++ sys/flexyadmin/models/file_manager.php
@@ -127,7 +127,7 @@ class File_manager Extends Model {
 				}
 				$names=filter_by($names,"_");
 				foreach($names as $name) {
-					unlink($this->map."/".$name);
+					if (file_exists($this->map."/".$name)) unlink($this->map."/".$name);
 				}
 				
 				/**
